qawithcsv/requirements.txt
langchain
langchain-community
langchain-openai
langchain-experimental
python-dotenv
pandas
streamlit
python-dotenv
tabulate